requirements.txt
dash-design-kit==1.6.2
dash==1.19.0
gunicorn==20.0.4
plotly==4.14.1
pandas==1.1.4
SQLAlchemy==1.3.20
psycopg2-binary==2.8.6
tenacity==6.2.0
dash_enterprise_auth==0.0.5
Werkzeug==2.0.0